8.lightrag-api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
tqdm>=4.65.0